% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lex.R
\name{lex}
\alias{lex}
\title{calculate lextale score}
\usage{
lex(answerVar)
}
\arguments{
\item{answerVar}{a tibble with ids,type (word/nonword),correct(0/1) and answer(0/1) variables}
}
\description{
to calculate lextale score, provided with a tibble of at least four variables
}
