% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lex_dutch.R
\name{lex_dutch}
\alias{lex_dutch}
\title{calculate lextale score, Dutch version}
\usage{
lex_dutch(dataframe)
}
\arguments{
\item{dataframe}{a binary values (0/1), at least 60 to returns one score}
}
\description{
to mark & calculate score for Dutch lextale-test, provided with answer variable (0/1)
}
