% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lextale-package.R
\docType{package}
\name{Lextale-package}
\alias{Lextale}
\alias{Lextale-package}
\title{Lextale: A Package to calculate scoring for LexTALE-test, English, German and Dutch versions.}
\description{
This package calculates the scoring for the English LexTALE-test if administered with the downloads using implementations that do not end with participants' score on the screen, e.g. online surveys.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Ghozayel/Lextale}
  \item Report bugs at \url{https://github.com/Ghozayel/Lextale/issues}
}

}
\author{
\strong{Maintainer}: Ghozayel Elotteebi \email{gzyl4work@gmail.com}

}
\keyword{internal}
