% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lextale.R
\name{lextale}
\alias{lextale}
\title{LexTALE \% correctAV}
\usage{
lextale(answerVar)
}
\arguments{
\item{var}{A lextale var must be provided as a numeric data, either 0 or 1, with 60 responses at least for each participant.}
}
\description{
This package calculates the \% correctAV scoring for the English LexTALE-test if administered with the downloads using implementations that do not end with participants' score on the screen, e.g. online surveys
}
\examples{
lextale(sample(c(0,1), replace = TRUE, 600))
}
