% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lex.R
\name{lex}
\alias{lex}
\title{calculate lextale score}
\usage{
lex(answerVar)
}
\arguments{
\item{answerVar}{a binary values (0/1), at least 60 to returns one score}
}
\description{
to calculate lextale score, provided with answer variable (0/1)
}
