% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lex.R
\name{lex}
\alias{lex}
\title{calculate lextale score}
\arguments{
\item{answerVar}{a 0/1 values responses to lextale test}
}
\description{
calculate the lextale score
}
\examples{
lex(sample(c(0,1), replace = TRUE, 600))
}
